+++ v-bro-repo-persistence-xml/src/main/java/com/tschanz/v_bro/repo/persistence/xml/idref_parser/XmlIdElementPosInfo.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class XmlIdElementPosInfo {
     @Getter private final String name;
     @Getter private final String elementId;
-    @Getter private final int startBytePos;
-    @Getter private final int endBytePos;
+    @Getter private final long startBytePos;
+    @Getter private final long endBytePos;
     @Getter private final List<String> idRefs;
 }
